requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
go-task-bin==3.49.1
greenlet==3.3.2
gunicorn==25.3.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
ruff==0.15.8
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.7
